Move VCS modules into larch.git package (packaging 2/9)
Moves git, gh, repo_roots, rebase, push, pr, pr_body, and merge from
the flat python/ tree into python/larch/git/. All importers updated to
use 'from larch.git import X' and cli.py registry updated to
'larch.git.X' module paths. Updates the subprocess, complexity, and
topology baselines to reflect the new file paths.
