+++ python/larch/git/pr.py
@@ -8,10 +8,10 @@ from dataclasses import dataclass
 
 import argparse
 import sys
-import gh
-import git
-import pr_body
-import push
+from larch.git import gh
+from larch.git import git
+from larch.git import pr_body
+from larch.git import push
 import tracking_issue
 from pathlib import Path
 from larch.errors import ShipError
